requirements.txt
Flask
flask-cors
flask-apscheduler
pandas
mysql-connector-python
requests
faker